Fixed the kick option in Session players not kicking the proper person, and kicking either the host/person operating the menu.
Also fixed the issue where when you chose a player in Session Player, it doesn't actually match the player that you see in the top left.

This causes issues when using any of the options as it caused to apply to the person you are on, and not the player you chose.